avoid clang warning
git-svn-id: https://svn.r-project.org/R/trunk@62477 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/summary.c
@@ -257,7 +257,7 @@ static Rboolean iprod(int *x, R_xlen_t n, double *value, Rboolean narm)
 	    return updated;
 	}
     }
-    *value = s;
+    *value = (double) s;
 
     return updated;
 }
